refactor(mcp): own typed RMCP connection adapter
