+++ crates/tracedecay-mcp/src/server/mod.rs
@@ -3,6 +3,7 @@
 mod connection;
 mod dispatch;
 mod protocol;
+mod rmcp;
 mod settlement;
 
 pub use crate::lifecycle::{
@@ -15,6 +16,11 @@ pub use dispatch::{
     McpDispatchParams, McpDispatchRequest, ToolCallParams, dispatch_is_independent_read,
 };
 pub use protocol::{McpMethod, classify_mcp_method, initialize_result, resources_list_result};
+pub use rmcp::{
+    RmcpConnectionAdapter, RmcpInitializeResponseDecorator, RmcpSelectedProjectResponseAuthority,
+    RmcpWorkDeliverySettlement, await_dispatch_with_cancellation, project_server_retired_error,
+    rmcp_response_result,
+};
 pub use settlement::{
     ApplicationCancellationRegistration, DispatchControl, DispatchControlRequest, DispatchFailure,
     DispatchSettlement, DispatchToolPolicy, PreparedDispatchControl, RetainedDispatchAuthority,
